fix bug in newly introduced removal of small segments, fix job submission for correspondences

+++ segmentation/segmentForPipeline.m
@@ -11,7 +11,7 @@ seg = uint16(seg{1,1});
 % Remove small objects (less than 9 voxels) from segmentation
 segMask = bwareaopen(seg, 9);
 affImposed = imimposemin(aff, segMask);
-seg = watershed(affImposed, 26);
+seg = uint16(watershed(affImposed, 26));
 
 % If folder does not exist, create it
 saveFolder = fileparts(saveFile);
